This is the updated I2C Communication files.
The I2C_Master.c is the code sitting on the BeagleBone, and will handle
the interim communication between the ROS instance and the Arduino.

The I2C_Slave is the code for the Arduinos. It will handle the interactions
with the actual sensors, servos, and whatever else.

I2C_Master.c is fairly stable, though testing it today I have noticed the
BeagleBone can get out of sync with the Arduino, which when that happens
all future communication fails. I'm not entirely sure of the specific cause.